Move test access_token generation to config-test
Also change the AUTH_AUDIENCE default to localhost
