+++ src/config/config-test.ts
@@ -1,6 +1,9 @@
+import * as auth from 'hapi-auth-jwt2';
 import { Container } from 'inversify';
+import { sign } from 'jsonwebtoken';
 import * as Knex from 'knex';
 
+import { AccessToken, jwtOptionsInjectSymbol, teamTokenClaimKey } from '../authentication';
 import AuthenticationModule from '../authentication/authentication-module';
 import { goodOptionsInjectSymbol } from '../server';
 import { fetchMock } from '../shared/fetch';
@@ -27,6 +30,44 @@ const charlesKnex = Knex({
   useNullAsDefault: true,
 });
 
+// Access token parameters
+export const audience = 'https://api.foo.com';
+export const issuer = 'https://issuer.foo.com';
+export const secretKey = 'shhhhhhh';
+export const algorithm = 'HS256';
+
+function getJwtOptions(): auth.JWTStrategyOptions {
+  return {
+    // Get the complete decoded token, because we need info from the header (the kid)
+    complete: true,
+    key: secretKey,
+    // Validate the audience, issuer, algorithm and expiration.
+    verifyOptions: {
+      audience,
+      issuer,
+      algorithms: [algorithm],
+      ignoreExpiration: false,
+    },
+  };
+}
+
+export function getAccessToken(sub: string, teamToken?: string, email?: string) {
+  let payload: Partial<AccessToken> = {
+    iss: issuer,
+    sub,
+    aud: [audience],
+    azp: 'azp',
+    scope: 'openid profile email',
+  };
+  if (teamToken) {
+    payload = { ...payload, [teamTokenClaimKey]: teamToken };
+  }
+  if (email) {
+    payload = { ...payload, email };
+  }
+  return sign(payload, secretKey);
+}
+
 export default (kernel: Container) => {
   developmentConfig(kernel);
   kernel.rebind(fetchInjectSymbol).toConstantValue(fetchMock.fetchMock);
@@ -34,4 +75,5 @@ export default (kernel: Container) => {
   kernel.rebind(GitlabClient.injectSymbol).toConstantValue(getClient());
   kernel.rebind(charlesKnexInjectSymbol).toConstantValue(charlesKnex);
   kernel.rebind(loggerInjectSymbol).toConstantValue(logger);
+  kernel.rebind(jwtOptionsInjectSymbol).toConstantValue(getJwtOptions());
 };
